requirements.txt
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.2.0
tensorflow>=2.12.0
rdkit-pypi>=2023.9.1
joblib>=1.2.0
logging
